Extract rail-connected load placement solver
